// frontend/src/client/types.gen.ts
export type Product = {
    name: string;
    description?: (string | null);
    price: number;
    quantity: number;
    category: string;
    id?: string;
};